Added newline 5

+++ trap1.cpp
@@ -353,7 +353,7 @@ void run_function(int func_id) {
         }
 
         if(func_id == 5 || func_id == 6|| func_id == 8 || func_id == 9){
-            printf("Value at %d\n", get_at_results_val[0].of.i64);
+            printf("Value at %d\n", get_at_results_val[0].of.i32);
         }
         printf("This is the actual value -> %d\n", it);
         printf("This is the 2nd actual value -> %d\n", bt);
